resource-timing/embed-reload-back.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8" />
<title>Resource Timing embed reload - back button navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    let t = async_test("Test that object reloads are not observable by the parent, even after history navigations");
    window.open("resources/embed-reload-back.sub.html");
    addEventListener("message", t.step_func(e => {
        assert_equals(e.data, "PASS");
        t.done();
    }));
</script>
